camsim

+++ scripts/cam_sim.py
@@ -41,8 +41,8 @@ class Cam_sim():
 					# image = bridge.imgmsg_to_cv2(image_msg, "bgr8")
 					# image = cv2.resize(image, (64, 64))
 					cmd = memory['position']
-					print (float(cmd.x)/5, ' ', float(cmd.y)/5)
-					self.images[float(cmd.x)/5][float(cmd.y)/5] = image
+					print (int(cmd.x)/5, ' ', int(cmd.y)/5)
+					self.images[int(cmd.x)/5][int(cmd.y)/5] = image
 
 					#x_vector.append(float(cmd.x))
 					#y_vector.append(float(cmd.y))
